Move Promises to API module
